Add load % as an input for port rate

+++ server/winpcapport.cpp
@@ -263,7 +263,7 @@ void WinPcapPort::populateInterfaceInfo()
     interfaceInfo_ = new InterfaceInfo;
 
     interfaceInfo_->speed = adapter->TransmitLinkSpeed != quint64(-1) ?
-        adapter->TransmitLinkSpeed/ulong(1e6) : 0;
+        adapter->TransmitLinkSpeed/1e6 : 0;
     interfaceInfo_->mtu = adapter->Mtu;
 
     if (adapter->PhysicalAddressLength == 6) {
